合并controller # Conflicts: #	src/main/java/cn/com/connext/oms/web/Controller/PageController.java #	src/main/resources/templates/pages/details/commodity/goods-list.html
