+++ src/main/resources/templates/pages/details/orders/order-list.html
@@ -62,7 +62,7 @@
                         <label>{{checkedDate}}</label>
                     </td>
                     <td>{{list.basicState}}</td>
-                    <td>{{list.channelCode}}</td>
+                    <td>{{list.orderCode}}</td>
                     <td>{{list.basicState}}</td>
                     <td>{{list.orderSource}}</td>
                     <td>{{list.paymentWay}}</td>
